refactor: move identity provider protocol to outside
