--- src/data/protocols/identity/index.ts
@@ -1,1 +1,0 @@
-export * from './identity-provider';
